# Optimize Memory Allocation with Object Pool