# Optimize Memory Allocation with Object Pool

## Changes
Implement object pooling for Message and bytes.Buffer to reduce memory allocations and GC pressure, improving performance.

### Key Changes
- Add `pool.go` implementing object pools for Message and bytes.Buffer
- Modify `writer.go` Send and SendData methods to use object pools
- Add benchmark tests to verify performance improvements

### Performance Improvements
| Scenario | Before | After | Improvement |
|----------|---------|--------|-------------|
| Message Allocation | 264.2 ns/op, 2 allocs/op | 8.506 ns/op, 0 allocs/op | ~31x |
| Buffer Allocation | 52.23 ns/op, 2 allocs/op | 9.439 ns/op, 0 allocs/op | ~5.5x |
| Message Encoding | 153.5 ns/op, 5 allocs/op | 138.5 ns/op, 3 allocs/op | ~10% |

## Testing
- [x] Added benchmark tests to verify performance improvements
- [x] Maintained existing functionality
- [x] Preserved SSE event name validation

## Impact
- Significantly reduced memory allocations
- Lowered GC pressure
- Improved message processing performance
- No functional changes, fully backward compatible